Ignore previous commit about Java passing arrays by value. This isn't correct -- it passes the VALUE of the reference. This means that if you pass an array into a method and modify its contents within that method, the modifications will persist (and be visible from the calling method) once the method returns. However, consider this:
public void myMethod(double[] myArray) {

  myArray = someOtherArray.clone();
}

Suppose myMethod() is called from a method in some other class. When
myMethod() returns, the contents of the array passed in to myMethod()
will NOT be the same as the contents of cloned array.

Changed TopicScore.java and WordScore.java to avoid cloning arrays
when sampling alpha and beta to avoid this kind of confusion.